完成Fbx动画播放功能基础版本，修复更新channels矩阵里的各种bug，主要以时间bug为主，修复glsl中layout ivec不正常问题，改为了vec4，但是这个问题还需要跟踪，解决动画最后一阵跳转不流畅问题

+++ Mesh.h
@@ -83,8 +83,8 @@ public:
 	int GetIndicesNum() const;
 
 	//----bone ids
-	void AddBoneId(GLuint id);
-	const GLuint* GetBoneIds() const;
+	void AddBoneId(GLfloat id);
+	const GLfloat* GetBoneIds() const;
 	int GetBoneIdsMemSize() const;
 	int GetBoneIdsNum() const;
 
@@ -119,7 +119,7 @@ private:
 	QVector<float>						m_tangents;
 	QVector<float>						m_binormals;
 	QVector<float>						m_uvs1;
-	QVector<GLuint>						m_boneIds;
+	QVector<float>						m_boneIds;
 	QVector<float>						m_boneWeights;
 	QVector<VertInfo>					m_vertInfoVec;
